+++ aether_ui_win32.c
@@ -2334,7 +2334,7 @@ void aether_ui_canvas_begin_path_impl(int canvas_id) {
     canvas_add_cmd(canvas_id, c);
 }
 
-void aether_ui_canvas_move_to_impl(int canvas_id, float x, float y) {
+void aether_ui_canvas_move_to_impl(int canvas_id, double x, double y) {
     CanvasCmd c = {0}; c.k = CV_MOVE; c.p0 = x; c.p1 = y;
     canvas_add_cmd(canvas_id, c);
     if (canvas_id >= 1 && canvas_id <= canvas_count) {
@@ -2343,7 +2343,7 @@ void aether_ui_canvas_move_to_impl(int canvas_id, float x, float y) {
     }
 }
 
-void aether_ui_canvas_line_to_impl(int canvas_id, float x, float y) {
+void aether_ui_canvas_line_to_impl(int canvas_id, double x, double y) {
     if (canvas_id < 1 || canvas_id > canvas_count) return;
     Canvas* cv = &canvases[canvas_id - 1];
     CanvasCmd c = {0}; c.k = CV_LINE;
@@ -2352,24 +2352,24 @@ void aether_ui_canvas_line_to_impl(int canvas_id, float x, float y) {
     cv->cur_x = x; cv->cur_y = y;
 }
 
-void aether_ui_canvas_stroke_impl(int canvas_id, float r, float g, float b,
-                                   float a, float line_width) {
+void aether_ui_canvas_stroke_impl(int canvas_id, double r, double g, double b,
+                                   double a, double line_width) {
     CanvasCmd c = {0};
     c.k = CV_STROKE; c.cr = r; c.cg = g; c.cb = b; c.calpha = a; c.p0 = line_width;
     canvas_add_cmd(canvas_id, c);
 }
 
-void aether_ui_canvas_fill_rect_impl(int canvas_id, float x, float y,
-                                      float w, float h,
-                                      float r, float g, float b, float a) {
+void aether_ui_canvas_fill_rect_impl(int canvas_id, double x, double y,
+                                      double w, double h,
+                                      double r, double g, double b, double a) {
     CanvasCmd c = {0};
     c.k = CV_FILL_RECT; c.p0 = x; c.p1 = y; c.p2 = w; c.p3 = h;
     c.cr = r; c.cg = g; c.cb = b; c.calpha = a;
     canvas_add_cmd(canvas_id, c);
 }
 
-void aether_ui_canvas_arc_impl(int canvas_id, float cx, float cy, float radius,
-                                float start_angle, float end_angle) {
+void aether_ui_canvas_arc_impl(int canvas_id, double cx, double cy, double radius,
+                                double start_angle, double end_angle) {
     CanvasCmd c = {0};
     c.k = CV_ARC; c.p0 = cx; c.p1 = cy; c.p2 = radius;
     c.a0 = start_angle; c.a1 = end_angle;
@@ -2381,15 +2381,15 @@ void aether_ui_canvas_close_path_impl(int canvas_id) {
     canvas_add_cmd(canvas_id, c);
 }
 
-void aether_ui_canvas_fill_impl(int canvas_id, float r, float g, float b, float a) {
+void aether_ui_canvas_fill_impl(int canvas_id, double r, double g, double b, double a) {
     CanvasCmd c = {0};
     c.k = CV_FILL; c.cr = r; c.cg = g; c.cb = b; c.calpha = a;
     canvas_add_cmd(canvas_id, c);
 }
 
 void aether_ui_canvas_fill_text_impl(int canvas_id, const char* text,
-                                      float x, float y, float font_size,
-                                      float r, float g, float b, float a) {
+                                      double x, double y, double font_size,
+                                      double r, double g, double b, double a) {
     CanvasCmd c = {0};
     c.k = CV_FILL_TEXT; c.p0 = x; c.p1 = y; c.p2 = font_size;
     c.cr = r; c.cg = g; c.cb = b; c.calpha = a;
@@ -2397,7 +2397,7 @@ void aether_ui_canvas_fill_text_impl(int canvas_id, const char* text,
     canvas_add_cmd(canvas_id, c);
 }
 
-void aether_ui_canvas_draw_image_impl(int canvas_id, float x, float y,
+void aether_ui_canvas_draw_image_impl(int canvas_id, double x, double y,
                                        int iw, int ih,
                                        const unsigned char* rgba, int byte_len) {
     if (iw <= 0 || ih <= 0 || !rgba) return;
@@ -2428,8 +2428,8 @@ static void win32_copy_stops(CanvasCmd* c, int n_stops,
 }
 
 void aether_ui_canvas_fill_linear_gradient_impl(int canvas_id,
-        float x1, float y1, float x2, float y2,
-        int n_stops, void* offsets, void* rgba, float line_width) {
+        double x1, double y1, double x2, double y2,
+        int n_stops, void* offsets, void* rgba, double line_width) {
     CanvasCmd c = {0};
     c.k = CV_FILL_LINEAR; c.gx1 = x1; c.gy1 = y1; c.gx2 = x2; c.gy2 = y2;
     c.grad_line_width = line_width;
@@ -2438,8 +2438,8 @@ void aether_ui_canvas_fill_linear_gradient_impl(int canvas_id,
 }
 
 void aether_ui_canvas_fill_radial_gradient_impl(int canvas_id,
-        float cx, float cy, float radius, float fx, float fy,
-        int n_stops, void* offsets, void* rgba, float line_width) {
+        double cx, double cy, double radius, double fx, double fy,
+        int n_stops, void* offsets, void* rgba, double line_width) {
     CanvasCmd c = {0};
     c.k = CV_FILL_RADIAL; c.gx1 = cx; c.gy1 = cy; c.gr = radius;
     c.gfx = fx; c.gfy = fy; c.grad_line_width = line_width;
